+++ src/remotes/zsimulator/zsimwebview/ulascreen.ts
@@ -41,11 +41,11 @@ export class UlaScreen {
 	 * @param imgData A reusable array to create the pixel data in.
 	 * @param ulaScreen The ULA screen data. Pixels + color attributes.
 	 * @param time An optional time in ms which is used for the flashing of the color attributes.
-	 * The flash frequency is 1.6Hz.
+	 * The flash frequency is 1/640ms.
 	 */
 	public static drawUlaScreen(ctx: CanvasRenderingContext2D, imgData: ImageData, ulaScreen: Uint8Array, time = 0) {
 		// Check time. Calculate remainder.
-		const interval = 625;	// 625 ms
+		const interval = 640;	// 640 ms
 		const remainder = time % interval;
 		const flash = (remainder >= interval / 2) ? 0x80 : 0; // 0x80 if colors should be exchanged